bug fix: добавила приложение api, переименовала приложение bid в orders
