--- hrspace/bid/validators.py
@@ -1,1 +1,0 @@
-SALARY_VALIDATOR: int = 1
