+++ hrspace/orders/validators.py
@@ -1,0 +1,1 @@
+SALARY_VALIDATOR: int = 1
